Initial commit: README and feature proposal files
This commit includes the initial version of the README file, which provides an overview of the project and instructions for running the application. It also includes some initial files for proposing a new feature, including a template for feature requests and a proposal checklist.

The goal of this commit is to provide a basic structure for the project and enable other developers to get up and running quickly. Further commits will include additional features, bug fixes, and improvements to the application.